+++ src/components/rsui-components.js
@@ -2,5 +2,6 @@ import BrandedNavbarMenu from './BrandedNavbarMenu.vue';
 import BrandedNavColumn from './BrandedNavColumn.vue';
 import ToggleNav from './ToggleNav.vue';
 import Accordion from './Accordion.vue';
+import Accordion from './TypeaheadSelect.vue';
 
-export const components = { BrandedNavbarMenu : BrandedNavbarMenu, BrandedNavColumn : BrandedNavColumn, ToggleNav : ToggleNav, Accordion : Accordion };
+export const components = { BrandedNavbarMenu : BrandedNavbarMenu, BrandedNavColumn : BrandedNavColumn, ToggleNav : ToggleNav, Accordion : Accordion, TypeaheadSelect : TypeaheadSelect };