+++ changelog.md
@@ -5,7 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.5.6] - 2020-17-9
+## [1.5.8] - 2021-5-26
+
+**Added**:
+
+- Added TypeaheadSelect component
+
+## [1.5.7] - 2020-9-9
+
+**Fixes**:
+
+- Fixes bug in styling on BrandedNavColumn component
+
+## [1.5.6] - 2020-7-17
 
 **Updated**: